+++ chain/types/ethtypes/eth_7702_env.go
@@ -1,0 +1,9 @@
+package ethtypes
+
+import "github.com/filecoin-project/go-address"
+
+// Stub declarations for builds without the eip7702_enabled tag so references compile.
+var EvmApplyAndCallActorAddr address.Address
+
+func applyEnvEvmApplyAndCallAddr() {}
+