+++ chain/types/ethtypes/eth_7702_featureflag_test.go
@@ -8,19 +8,19 @@ import (
     "github.com/stretchr/testify/require"
 
     "github.com/filecoin-project/go-address"
+    "github.com/filecoin-project/go-state-types/abi"
     "github.com/filecoin-project/go-state-types/big"
     "github.com/filecoin-project/lotus/build/buildconstants"
-    delegator "github.com/filecoin-project/lotus/chain/actors/builtin/delegator"
 )
 
 // Validates that when the eip7702_enabled build tag is set and the DelegatorActorAddr
 // is configured, ToUnsignedFilecoinMessage constructs a message targeting the Delegator
 // actor with CBOR-encoded params.
 func Test7702_ToUnsignedFilecoinMessage_FeatureFlag(t *testing.T) {
-    // Configure a fake Delegator actor address
+    // Configure a fake EVM ApplyAndCall actor address
     a, err := address.NewIDAddress(1234)
     require.NoError(t, err)
-    DelegatorActorAddr = a
+    EvmApplyAndCallActorAddr = a
 
     // Minimal 7702 tx with one authorization
     var to EthAddress
@@ -43,7 +43,7 @@ func Test7702_ToUnsignedFilecoinMessage_FeatureFlag(t *testing.T) {
     require.NoError(t, err)
     msg, err := tx.ToUnsignedFilecoinMessageAtomic(from)
     require.NoError(t, err)
-    require.Equal(t, DelegatorActorAddr, msg.To)
-    require.EqualValues(t, delegator.MethodApplyAndCall, msg.Method)
+    require.Equal(t, EvmApplyAndCallActorAddr, msg.To)
+    require.EqualValues(t, abi.MethodNum(MethodHash("ApplyAndCall")), msg.Method)
     require.NotEmpty(t, msg.Params)
 }